Make it build. Still need to determine how proc args and local vars will be stored and referenced by IR instructions.

+++ src/ast.h
@@ -674,7 +674,7 @@ struct Symbol {
 
     // TODO: Cleanup this struct.
     // NOTE: For SYMBOL_VAR
-    s64 offset;
+    s32 offset;
 };
 
 Symbol* new_symbol_decl(Allocator* allocator, SymbolKind kind, const char* name, Decl* decl);
